Make this plugin compatible with Node.js 10+ (#82)
Description

The zmq module is unmaintained for 3 years now, and it's no longer compatible with the latest versions of Node.js
However, a fork of that module has been made a year ago to solve that issue and, since then, the new module grew and is now the new standard to access 0mq with Node.js: zeromq

This PR replaces the zmq module with zeromq. Because it comes from a fork, the API is the same, so the changes are minimal.