Remove the deprecated notifications state (#84)
Real-time notifications' state value (can be either pending or done) is being removed from Kuzzle v2, so we need to update the cluster as well.

Note: functional tests will fail because this change is not yet in Kuzzle v2 (among other things). This PR can still be merged if only the linter and unit tests pass.